+++ src/js/page/odoo/commands/backoffice/lang.mjs
@@ -152,6 +152,7 @@ function getOptions(this: Terminal, arg_name: string): Promise<Array<string>> {
       [],
       ['name'],
       this.getContext({active_test: true}),
+      {orderBy: 'name ASC'},
       item => item.name,
     );
   } else if (arg_name === 'lang') {
@@ -161,6 +162,7 @@ function getOptions(this: Terminal, arg_name: string): Promise<Array<string>> {
       [],
       ['code'],
       this.getContext({active_test: true}),
+      {orderBy: 'code ASC'},
       item => item.code,
     );
   }