+++ src/js/page/odoo/commands/backoffice/settings.mjs
@@ -31,6 +31,7 @@ function getOptions(this: Terminal, arg_name: string) {
       [],
       ['name'],
       this.getContext({active_test: true}),
+      {orderBy: 'name ASC'},
       item => item.name,
     );
   }