+++ src/js/page/odoo/commands/common/jstest.mjs
@@ -30,6 +30,7 @@ function getOptions(this: Terminal, arg_name: string) {
       [],
       ['name'],
       this.getContext({active_test: true}),
+      {orderBy: 'name ASC'},
       item => item.name,
     );
   }