+++ src/js/page/odoo/commands/common/read.mjs
@@ -61,6 +61,7 @@ function getOptions(this: Terminal, arg_name: string) {
       [],
       ['model'],
       this.getContext({active_test: true}),
+      {orderBy: 'model ASC'},
       item => item.model,
     );
   }