+++ src/js/page/odoo/commands/common/upgrade.mjs
@@ -56,6 +56,7 @@ function getOptions(this: Terminal, arg_name: string) {
       [],
       ['name'],
       this.getContext({active_test: true}),
+      {orderBy: 'name ASC'},
       item => item.name,
     );
   }