+++ src/js/page/terminal/libs/graphics/__all__.mjs
@@ -6,6 +6,7 @@ import funcDDestroyWindow from './2d_destroy_window';
 import func2DLine from './2d_line';
 import func2DRect from './2d_rect';
 import func2DClear from './2d_clear';
+import func2DText from './2d_text';
 import type VMachine from '@trash/vmachine';
 
 export default function (vm: VMachine) {
@@ -14,4 +15,5 @@ export default function (vm: VMachine) {
   vm.registerCommand('2d_line', func2DLine());
   vm.registerCommand('2d_rect', func2DRect());
   vm.registerCommand('2d_clear', func2DClear());
+  vm.registerCommand('2d_text', func2DText());
 }